requirements.txt
pillow
django
djangorestframework
django-cors-headers
django-multiselectfield
